move style sheet editor to ide
